Move bolding into \recommend
Recommendations using the `\recommend` text were being omitted entirely.
Only the bold colon was showing.

Reversing the order of the text and the call to `\addcontentsline` got
the text to show up in place at least, but it failed to add the content
to the TOC.

This appears to be a regression in Pandoc. (Or we were abusing it, and
got busted. One of those two.)

Running pandoc to emit tex instead of pdf showed no calls to
`\recommend`, only to `\textbf`. That command to set text as boldface
was being injected by the surrounding bold Markdown. It appears that the
text put within the command does not like including other commands in
the latest Pandoc.

In order to work around this, shift the formatting we'd been doing by
hand into the `\recommend` command itself. This includes bolding the
text and the following colon.

Existing uses have been updated to match in the audit starter.

Tested against:

pandoc 2.3.1
Compiled with pandoc-types 1.17.5.1, texmath 0.11.1.1, skylighting 0.7.3

tlmgr revision 48598 (2018-09-06 17:01:34 +0200)
tlmgr using installation: /usr/local/texlive/2018basic
TeX Live (http://tug.org/texlive) version 2018